refactor: relocate hooks folder out of components and remove unused react query code
