--- client/src/components/hooks/card_harness.tsx
@@ -1,24 +1,0 @@
-import { useCard } from "./use_card";
-
-function CardHarness({ id }: { id: number }) {
-  const query = useCard(id);
-
-  return (
-    <div>
-      <p>{query.isPending ? "isPending" : null}</p>
-
-      <p>{query.status}</p>
-
-      <p>{query.isError ? "isError" : null}</p>
-      {query.error && <p>Error: {query.error.message}</p>}
-
-      <p>{query.isSuccess ? "isSuccess" : null}</p>
-
-      {query.data && <p>haveData</p>}
-
-      {query.data && <p>{query.data.name}</p>}
-    </div>
-  );
-}
-
-export { CardHarness };
