--- client/src/components/hooks/mock_query_provider.tsx
@@ -1,18 +1,0 @@
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { PropsWithChildren } from "react";
-
-function MockQueryProvider({ children }: PropsWithChildren) {
-  const mockClient = new QueryClient({
-    defaultOptions: {
-      queries: {
-        retry: false,
-      },
-    },
-  });
-
-  return (
-    <QueryClientProvider client={mockClient}>{children}</QueryClientProvider>
-  );
-}
-
-export { MockQueryProvider };
